Removing unneeded code
