+++ .gitignore
@@ -1,3 +1,5 @@
+.vscode
+
 shit
 local_config
