--- src/other/addrs_Bloom/.gitignore
@@ -1,2 +1,0 @@
-file*
-x*
