--- src/other/addrs_Bloom/shitem
@@ -1,6 +1,0 @@
-bin/addrs
-ls -l output-final/blooms | cut -c28-36,50-100 | sed 's/ 00/|00/' | sed 's/.bloom//' | sed 's/ //g' | sort -n | tr '|' '\t' | grep -v "^$" | awk '{print $2 "\t" $1}' | sort >file.1
-ls -l output-final/blooms.old | cut -c28-36,50-100 | sed 's/ 00/|00/' | sed 's/.blooms.bin//' | sed 's/ //g' | sort -n | tr '|' '\t' | grep -v "^$" | awk '{print $2 "\t" $1}' | sort >file.2
-ls -l output-zips | cut -c28-35,49-100 | sed 's/ 00/|00/' | sed 's/.bloom.gz//' | sed 's/ //g' | sort -n | tr '|' '\t' | grep -v "^$" | awk '{print $2 "\t" $1}' | sort >file.3
-#join file.2 file.3 >file.4
-join file.1 file.3 | sed 's/ /,/g' | cut -f2 -d'-' >x.csv
